Stop prometheus before update clients list
Stopping prometheus before update clients list to have consistent data.
Here is the case:

1. One deploys prometheus with N clients
2. One add a client node to prometheus installation
3. New node gets ports from existing one, old one gets new ports (ports
   numbers depends on position of the node in the list)
4. Node configuration gets updated
5. Autossh on changed nodes restarted. As we have a new node, its
   autossh restarted definitely. It will have ports which previous node
   had had before
6. Node sends info to prometheus server. As long as job name is the same
   in prometheus, it pushes info from new host with old job name, as job
   name was not updated yet
7. Client nodes configured
8. Prometheus server part started configuration
9. Prometheus server updates its config and restarts. Now it contains
   new job name for the new node (job name depends on node ansible_host
   var value).

After all of these there will me some time (about 2 minutes by my
measures) while prometheus will store data for the job name of one host
with the data of another one. It will break some dashboards which are
based on the job name (there are a lot of them!!). To avoid such a
situation, stop prometheus server container before reconfigure clients.
It will definitely lead to small gaps in prometheus stability but for
now it is the best we have.